+++ src/http/controllers/register.ts
@@ -1,0 +1,24 @@
+import { prisma } from "@/lib/prisma";
+import { FastifyReply, FastifyRequest } from "fastify";
+import { z } from "zod";
+
+export async function register(request: FastifyRequest, response: FastifyReply) {
+
+    const registryBodySchema = z.object({
+        name: z.string(),
+        email: z.string().email(),
+        password: z.string().min(6),
+    });
+
+    const { name, email, password } = registryBodySchema.parse(request.body);
+
+    await prisma.user.create({
+        data: {
+            nome: name,
+            email,
+            password_hash: password
+        }
+    });
+
+    return response.status(201).send();
+}
